A little refactoring, commenting and documenting ideas behind this specific implementation changes
