+++ Weather/Modules/CityDetail/Builder/WeatherDetailDefaultBuilder.swift
@@ -13,5 +13,4 @@ struct WeatherDetailDefaultBuilder : WeatherDetailBuilder {
 
         return view       
     }
-    
 }
