+++ Weather/Modules/CityDetail/Core/Interactor/WeatherDetailDefaultInteractor.swift
@@ -19,5 +19,4 @@ class WeatherDetailDefaultInteractor: WeatherDetailInteractor {
             }
         }
     }
-    
 }
