+++ Weather/Modules/CityDetail/Core/Presenter/WeatherDetailDefaultPresenter.swift
@@ -48,5 +48,4 @@ class WeatherDetailDefaultPresenter: WeatherDetailPresenter {
                                       temperature: data.temperature + data.temperatureUnit,
                                       forecasts: forecasts)
     }
-    
 }
