+++ Weather/Modules/CityDetail/Core/View/WeatherDetailViewController.swift
@@ -39,5 +39,4 @@ class WeatherDetailViewController: UIViewController, WeatherDetailView {
         
         self.presentViewController(alert, animated: true, completion: nil)
     }
-
 }
