+++ Weather/Modules/CityDetail/Core/View/WeatherForecastView.swift
@@ -4,5 +4,4 @@ class WeatherForecastView: UIView {
     
     @IBOutlet weak var dayLabel: UILabel!
     @IBOutlet weak var tempLabel: UILabel!
-
 }
